Add files via upload
his code is performing data preprocessing, data visualization, feature selection, and classification tasks 
on a dataset. The dataset is loaded and merged with another dataset, then duplicates are dropped, and 
non-numeric columns are encoded. The dataset is then scaled, and histograms, density plots, and 
boxplots are generated for both the original and scaled datasets. A correlation matrix is computed to 
identify relevant features for the classification task. Relevant features are selected and a K-Nearest 
Neighbors (KNN) classifier is trained using GridSearchCV to tune the hyperparameters. Finally, the model 
is evaluated using accuracy, precision, recall, and F1 score.

The code uses several Python libraries, including pandas, numpy, sklearn, matplotlib, seaborn, and 
KMeans. The dataset used in the code is loaded from two files: "Dataset.txt" and "Attack_types.txt". The 
fisst file contains the original dataset, while the second file contains information about attack types.
Merges the two datasets and drops duplicate rows. Encodes non-numeric columns using the Label 
Encoder function. Scales the dataset using the StandardScaler function. Plots histograms and density 
plots of the original and scaled datasets. Plots box plots of the original and scaled datasets. Calculates 
the correlation matrix of the scaled dataset and plots it as a heatmap Selects relevant features using a 
correlation threshold of 0.4 with the target variable. Splits the dataset into training and testing sets.
Trains a KNN classifier using a grid search for hyper parameter tuning. Evaluates the performance of the 
KNN classifier on the testing set by calculating accuracy, precision, recall, and F1 score